// 02 - M01 Search ME/model/NorthwindProduct.ts
export interface NorthwindProduct {
    ProductID: number;
    ProductName: string;
    SupplierID: number;
    CategoryID: number;
    QuantityPerUnit: string;
    UnitPrice: number;
    UnitsInStock: string;
    UnitsOnOrder: number;
    ReorderLevel: number;
    Discontinued: boolean;
}